% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_large_polygons.R
\name{amend_large_polygons}
\alias{amend_large_polygons}
\title{Extract large polygons and assign additional layers}
\usage{
amend_large_polygons(vri_bem, glaciers, lakes, wetlands, bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}

\item{glaciers}{sf}

\item{lakes}{sf}

\item{wetlands}{sf}

\item{bem}{sf}
}
\description{
Extract large polygons and adjacent polygons,
Cut in CEF glaciers and FWA wetlands and lakes
Merge BEM and correct assigned BEU/SDEC
}
