% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beu_corrections.R
\name{BEUMC_adjustment}
\alias{BEUMC_adjustment}
\title{BEU fix}
\usage{
BEUMC_adjustment(vri_bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}
}
\description{
Fix BEUs which are in wrong location
}
\details{
If BEUMC_S1 is missing, fill with BEUMC_S2 or BEUMC_S3 if they exist
If BEUMC are adjusted, SDEC also adjusted
}
