% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_db.R, R/read_layers.R
\name{cache_dir}
\alias{cache_dir}
\alias{cache_remove}
\title{Cache directory}
\usage{
cache_dir()

cache_remove()

cache_dir()

cache_remove()
}
\description{
Cache directory

Remove cache

Cache directory

Remove cache
}
