% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_capability_rating.R
\name{calc_capability_rating}
\alias{calc_capability_rating}
\title{calc_capability_rating}
\usage{
calc_capability_rating(rrm_dt, animal)
}
\arguments{
\item{rrm_dt}{data.table object that contains the rrm}

\item{animal}{character, "bear" or "moose"}
}
\value{
data.table rrm with new column capability_rating
}
\description{
calculate capability rating which is the best suitability rating among ecosystem regardless of structural stage
}
