% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hem_fields.R
\name{calc_hem_fields}
\alias{calc_hem_fields}
\title{Calculate all hem fields for the vri_bem database}
\usage{
calc_hem_fields(
  vri_bem,
  fire,
  hem_fields = SSGBM.VRI.BEM::hem_fields,
  current_year = year(Sys.Date())
)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{fire}{sf object that represent}

\item{hem_fields}{data.table containing a column called 'names' with the name of the hem fields to compute, default to the hem_field object in the package}

\item{curent_year}{numeric of the year use to compute age, default to year of current date}
}
\value{
sf object
}
\description{
Calculate all hem fields for the vri_bem database
}
