% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_vri.R
\name{check_allowed_bec_beu}
\alias{check_allowed_bec_beu}
\title{Check allowed BEC/BEU combinations}
\usage{
check_allowed_bec_beu(vri_bem, beu_bec)
}
\arguments{
\item{vri_bem}{sf object that represent the combined vri & bem polygon feature class}

\item{beu_bec}{data.table object of allowed BEC and BEM Code Combos}
}
\value{
sf object which contains adjusted map codes
}
\description{
This function uses a table of allowed BEC/BEU combinations and updates BEUs which are identified as mismatches.
}
\details{
This function checks all BEC/BEU combinations.
In cases where BEUs are assigned to BEC units they can't occur in, they will be updated to the correct BEU.
Note that mismatches may still occur if they aren't reflected in the allowed BEC/BEU code combos. Make sure to update regularly.
}
