% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vri_bem_output.R
\name{clean_vri_bem_output}
\alias{clean_vri_bem_output}
\title{Clean up the VRI-BEM-WHR output
This function cleans up the final VRI-BEM-WHR output (removes extraneous fields), removes slivers < 1m^2, and adds the unique ID (PolyID) and ECO_TYPE}
\usage{
clean_vri_bem_output(vri_bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}
}
\value{
"cleaned" vri_bem
}
\description{
Clean up the VRI-BEM-WHR output
This function cleans up the final VRI-BEM-WHR output (removes extraneous fields), removes slivers < 1m^2, and adds the unique ID (PolyID) and ECO_TYPE
}
