% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beu_from_rules_dt.R
\name{convert_rule_list_to_string_vector}
\alias{convert_rule_list_to_string_vector}
\title{convert_rule_list_to_string_vector}
\usage{
convert_rule_list_to_string_vector(rule_list)
}
\arguments{
\item{rule_list}{string list of values}
}
\value{
string
}
\description{
create string expression that convert list separate by commas to vector
}
\keyword{internal}
