% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_wetland.R
\name{correct_bem_from_wetlands}
\alias{correct_bem_from_wetlands}
\title{Correct BEM attributes based on Wetlands}
\usage{
correct_bem_from_wetlands(vri_bem, buc)
}
\arguments{
\item{vri_bem}{sf object that represents the VRI-BEM feature class}

\item{buc}{data.table object that represents BEU wetland update table}
}
\value{
sf object
}
\description{
Correct BEM attributes based on Wetlands
}
