% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beu_corrections.R
\name{correct_small_lakes}
\alias{correct_small_lakes}
\title{Fix missing lake polygons}
\usage{
correct_small_lakes(vri_bem, lakes)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}

\item{lakes}{sf}
}
\description{
Compare upland/non-lake areas with FWA Lakes and cut in lakes when necessary
}
