% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_forest_structure_variables.R
\name{create_empty_forest_structure_variables}
\alias{create_empty_forest_structure_variables}
\title{Create empty variables for structure of forest}
\usage{
create_empty_forest_structure_variables(unique_ecosystem_dt)
}
\arguments{
\item{unique_ecosystem_dt}{data.table of the containing information for each unique ecosystems}
}
\value{
data.table that contains the frequency of each unique ecosystem and generates the following empty column to be feed later on:
\itemize{
\item Forested (Y/N),
\item Strict_Climax,
\item Stand_Climax,
\item Stand_Age_0-15, Stand_Age_16-30, Stand_Age_31-50, Stand_Age_51-80, Stand_Age_80+,
\item Struct_Age_0-3, Struct_Age_4-10, Struct_Age_11-30, Struct_Age_31-40, Struct_Age_41-60, Struct_Age_61-80, Struct_Age_81-139, Struct_Age_140_249, Struct_Age_250+,
\item Snow_code
}
}
\description{
Generates a data of unique BGC label and habitat combinations for the Look up Table starting point
}
