% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hem_rasters.R
\name{create_hem_rasters}
\alias{create_hem_rasters}
\title{Create raster with layers for each hem fields}
\usage{
create_hem_rasters(
  vri_bem,
  path,
  xmin = 1020387.5,
  xmax = 1030387.5,
  ymin = 960987.5,
  ymax = 970987.5,
  resolution = 100,
  crs =
    "PROJCS[\\"NAD_1983_BC_Environment_Albers\\",GEOGCS[\\"GCS_North_American_1983\\",DATUM[\\"D_North_American_1983\\",SPHEROID[\\"GRS_1980\\",6378137.0,298.257222101]],PRIMEM[\\"Greenwich\\",0.0],UNIT[\\"Degree\\",0.0174532925199433]],PROJECTION[\\"Albers\\"],PARAMETER[\\"False_Easting\\",1000000.0],PARAMETER[\\"False_Northing\\",0.0],PARAMETER[\\"Central_Meridian\\",-126.0],PARAMETER[\\"Standard_Parallel_1\\",50.0],PARAMETER[\\"Standard_Parallel_2\\",58.5],PARAMETER[\\"Latitude_Of_Origin\\",45.0],UNIT[\\"Meter\\",1.0]]"
)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{path}{character string that represent the path the output raster will be save to}

\item{xmin}{numeric that represent the xmin of the extent of the raster}

\item{xmax}{numeric that represent the xmax of the extent of the raster}

\item{ymin}{numeric that represent the ymin of the extent of the raster}

\item{ymax}{numeric that represent the ymax of the extent of the raster}

\item{resolution}{numeric that represent the resolution of the raster}

\item{crs}{character string of the WKT to be used for the projection of the raster}
}
\value{
spatRaster object
}
\description{
Create raster with layers for each hem fields
}
