% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RRM_ecosystem.R
\name{create_RRM_ecosystem_huckleberry}
\alias{create_RRM_ecosystem_huckleberry}
\title{Create RRM ecosystem information for huckleberry}
\usage{
create_RRM_ecosystem_huckleberry(vri_bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}
}
\value{
Summary of Area by unique ecosystem
}
\description{
This function generates a unique list of ecosystems based on select huckleberry habitat characteristics.
It only includes existing (not potential) ecosystems
}
\details{
For each of the following combination compute the sum of the area
\itemize{
\item ECO_SEC
\item BGC_ZONE
\item BGC_SUBZON
\item BGC_VRT
\item BGC_PHASE
\item HUCK_ASP
\item HUCK_ELEV_Thold
\item CROWN_ALL (CROWN_ALL_1, CROWN_ALL_2, CROWN_ALL_3)
\item STRCT (STRCT_S1, STRCT_S2, STRCT_S3)
\item STAND (STAND_A1, STAND_A2, STAND_A3)
\item FORESTED (FORESTED_1, FORESTED_2, FORESTED_3)
}
}
