% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_ecosystem_dt.R
\name{create_unique_ecosytem_dt}
\alias{create_unique_ecosytem_dt}
\title{Create data of Unique ecosystem}
\usage{
create_unique_ecosytem_dt(ifc)
}
\arguments{
\item{ifc}{sf object that represent BEM (broad ecosystem mapping) features}
}
\value{
data.table
}
\description{
generates a data of unique BGC label and habitat combinations for the Look up Table starting point
}
