% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_updated_vri_bem.R
\name{create_updated_vri_bem}
\alias{create_updated_vri_bem}
\title{Creates the updated vri-bem object}
\usage{
create_updated_vri_bem(
  vri,
  bem,
  rivers,
  wetlands,
  beu_bec_csv,
  beu_wetland_update_csv,
  rules_dt,
  clear_site_ma = TRUE,
  use_ifelse = TRUE,
  return_intersection_dt = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vri}{sf object of vri}

\item{bem}{sf objects of bem}

\item{rivers}{sf object of rivers}

\item{wetlands}{sf object of wetlands}

\item{beu_bec_csv}{data.table of beu_bec_csv}

\item{beu_wetland_update_csv}{data.table of wetlands correction csv}

\item{rules_dt}{data.table of rules for improvement of beu}

\item{clear_site_ma}{boolean, if TRUE variable SITE_M1A, SITE_M2A will be cleared}

\item{use_ifelse}{boolean, if TRUE correction done after the combine_duplicated_BEUMC will only be applied on rows that were not affected by the correction of duplicated BEUMC}

\item{return_intersection_dt}{Boolean, if TRUE will return a list that contains the sf object and the intersection data.table of VRI and BEM}

\item{verbose}{boolean , if TRUE function will return message to indicate progress throughout the function execution}
}
\value{
sf object updated vri-bem
}
\description{
Creates the updated vri-bem object
}
