% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_raster_shelter_disturbance.R
\name{depth_calculations}
\alias{depth_calculations}
\title{Disturbance around shelter}
\usage{
depth_calculations(
  shelter,
  disturbance,
  depth = units::as_units("400 m"),
  rasterize = TRUE,
  ...
)
}
\arguments{
\item{shelter}{vri-bem sf object.}

\item{disturbance}{sf object}

\item{depth}{unit}

\item{rasterize}{A logical. Should the results be rasterized.}

\item{...}{}
}
\description{
Disturbance around shelter
}
