% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erase_geometry}
\alias{erase_geometry}
\title{Erase geometry from target sf object}
\usage{
erase_geometry(target, erase)
}
\arguments{
\item{target}{sf object}

\item{erase}{sf object}
}
\value{
sf object
}
\description{
Erase geometry from target sf object
}
