% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filtered_views}
\alias{filtered_views}
\title{Create the filtered views from the source table for the analysis}
\usage{
filtered_views(
  conn = init_conn(),
  wkt_filter,
  tables = c("BEM", "BURN", "CCB", "FIRE", "GLACIERS", "LAKES", "RIVERS", "VRI",
    "WETLANDS")
)
}
\arguments{
\item{conn}{A database connection.}

\item{wkt_filter}{A well-known text geometry of the area of interest.}

\item{tables}{A character vector of tables to create views from.}
}
\description{
Create the filtered views from the source table for the analysis
}
