% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_crown_area_dominant_values.R
\name{find_crown_area_dominant_values}
\alias{find_crown_area_dominant_values}
\title{Find Area-Dominant CROWN_BEAR and CROWN_MOOSE}
\usage{
find_crown_area_dominant_values(vri)
}
\arguments{
\item{vri}{sf object that represent VRI (vegetation resources inventory) features}
}
\value{
sf object containing VRI-BEM
}
\description{
This function determines the area-dominant CROWN_BEAR and CROWN_MOOSE values for each VRI-BEM polygon.
}
\details{
updated to assign crown values directly from the VRI. It's not necessary to intersect with the BEM
}
