% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raster_extract.R
\name{format_raster_extract}
\alias{format_raster_extract}
\title{Format the extraction of a raster}
\usage{
format_raster_extract(
  input_dt,
  character_attributes = NULL,
  date_attributes = NULL,
  factor_conv_list
)
}
\arguments{
\item{input_dt}{data.table extracted raster converted to data.table}

\item{character_attributes}{character vector that represent the attributes of type character that need to be transformed}

\item{date_attributes}{character vector that represent the attributes of type date that need to be transformed}

\item{factor_conv_list}{list of data.table that contains corresponding factor values for each variable that is not numeric}
}
\value{
data.table
}
\description{
Format the extraction of a raster
}
