% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rrm_dt.R
\name{format_rrm_dt}
\alias{format_rrm_dt}
\title{format_rrm_dt}
\usage{
format_rrm_dt(rrm_dt, animal)
}
\arguments{
\item{rrm_dt}{data.table object that contains the rrm}

\item{animal}{character, "bear" or "moose"}
}
\value{
data.table  formatted rrm_dt
}
\description{
format the rrm data to match the type of columns of vri-bem
}
