% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_unique_ecosystem_dt.R
\name{format_unique_ecosystem_dt}
\alias{format_unique_ecosystem_dt}
\title{Format unique ecosystem data}
\usage{
format_unique_ecosystem_dt(unique_ecosystem_dt)
}
\arguments{
\item{unique_ecosystem_dt}{data.table object that represent the unique ecosystems}
}
\value{
data.table
}
\description{
Format the unique ecosystem data, to make sure that any conversion that happened when reading the csv did not affect the expected class of each fields
}
