% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raster_extract.R
\name{format_vri_raster_extract}
\alias{format_vri_raster_extract}
\title{Format the extraction of a the VRI raster}
\usage{
format_vri_raster_extract(
  input_dt,
  character_attributes = c(paste0("BCLCS_LEVEL_", 1:5), paste0("SPECIES_CD_", 1:6),
    "LAND_COVER_CLASS_CD_1", "LINE_5_VEGETATION_COVER"),
  date_attributes = NULL
)
}
\arguments{
\item{input_dt}{data.table extracted raster converted to data.table}

\item{character_attributes}{character vector that represent the attributes of type character that need to be transformed}

\item{date_attributes}{character vector that represent the attributes of type date that need to be transformed}
}
\value{
data.table
}
\description{
Format the extraction of a the VRI raster
}
