% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve_beu.R
\name{improve_beu_from_rule_dt}
\alias{improve_beu_from_rule_dt}
\title{improve beu from rules}
\usage{
improve_beu_from_rule_dt(vri_bem, rules_dt)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{rules_dt}{data.table object that contains rule to apply to vri_bem to update beumc}
}
\value{
sf object
}
\description{
update beumc values based on rules
}
