% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bem_on_vri.R
\name{merge_bem_on_vri.data.table}
\alias{merge_bem_on_vri.data.table}
\title{Merge BEM attributes on VRI features}
\usage{
merge_bem_on_vri.data.table(vri, bem)
}
\arguments{
\item{vri}{data.table object that represent VRI (vegetation resources inventory) features}

\item{bem}{data.table object that represent BEM (broad ecosystem mapping) features}
}
\value{
data.table object that represent the original VRI with merged BEM attributes.
}
\description{
This function copies all of the attributes in the specified BEM (broad ecosystem mapping) to each polygon in the specified VRI (vegetation ressource inventory).
}
\details{
Its attribute table will have all of the BEM attribute fields added to the existing VRI attributes.
The BEM attribute fields will be populated by copying the BEM values.
}
