% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ccb_on_vri.R
\name{merge_ccb_on_vri}
\alias{merge_ccb_on_vri}
\title{Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory) features}
\usage{
merge_ccb_on_vri(ccb, vri)
}
\arguments{
\item{ccb}{sf object that represent CCB (consolidated cutblock) features}

\item{vri}{sf object that represent VRI (vegetation ressource inventory) features}
}
\value{
sf object that represent the intersections of ccb and vri
}
\description{
Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory) features
}
