% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_elevation_raster_on_bem.R
\name{merge_elevation_raster_on_bem}
\alias{merge_elevation_raster_on_bem}
\title{Compute slope and aspect from elevation and compute mean by bem polygon}
\usage{
merge_elevation_raster_on_bem(elev_raster, bem)
}
\arguments{
\item{elev_raster}{SpatRaster object that represent the elevation}

\item{bem}{sf object that represent BEM (broad ecosystem mapping) features}
}
\value{
data.table
}
\description{
Compute slope and aspect from elevation and compute mean by bem polygon
}
