% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_elevation_raster_on_bem.R
\name{merge_elevation_raster_on_sf}
\alias{merge_elevation_raster_on_sf}
\title{Compute slope and aspect from elevation and compute mean by ifc polygon}
\usage{
merge_elevation_raster_on_sf(elev_raster, ifc, elevation_threshold = 1400)
}
\arguments{
\item{elev_raster}{SpatRaster object that represent the elevation}

\item{ifc}{sf object}

\item{elevation_threshold}{numeric elevation threshold use to create above elevation indicator}
}
\value{
data.table
}
\description{
Compute slope and aspect from elevation and compute mean by ifc polygon
}
