% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fire_perimeters.R
\name{merge_fire_perimeters}
\alias{merge_fire_perimeters}
\title{Add fire perimeters}
\usage{
merge_fire_perimeters(vri_bem, fire_perimeter)
}
\arguments{
\item{vri_bem}{sf object}

\item{fire_perimeter}{sf object that represents historical BC Wildfire Fire Perimeters features}
}
\value{
sf object that represent the intersections of vri_bem and fire_perimeter
}
\description{
Joins historic fire perimeters to vri_bem by percent overlap
}
