% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_geometry}
\alias{merge_geometry}
\title{Union merge between CCB attributes and VRI features}
\usage{
merge_geometry(x, y, tolerance = units::as_units("100 m2"), label = NULL)
}
\arguments{
\item{x}{sf object}

\item{y}{sf object}
}
\value{
sf object that represent the intersections of x and y
}
\description{
Performs a Union merge between y attributes on x features
}
