% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rrm_on_vri.R
\name{merge_rrm_on_vri_moose}
\alias{merge_rrm_on_vri_moose}
\title{merge_rrm_on_vri_moose}
\usage{
merge_rrm_on_vri_moose(vri_bem, rrm_dt, return_sf = TRUE)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{rrm_dt}{data.table object that contains the rrm}

\item{return_sf}{logical, if TRUE  return sf object , if FALSE return data.table object and update by reference}
}
\value{
sf  vri-bem object with new columns for rating
}
\description{
merge suitability and capability rating from rrm onto vri-bem
and calculate highest value and weighted average for each scores
}
