% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_unique_ecosystem_fields.R
\name{merge_unique_ecosystem_fields}
\alias{merge_unique_ecosystem_fields}
\title{Assign values from unique ecosystems table}
\usage{
merge_unique_ecosystem_fields(ifc, unique_ecosystem_dt)
}
\arguments{
\item{ifc}{sf object}
}
\value{
sf object
}
\description{
Populates the TEIS fields, REALM, GROUP, CLASS, and KIND as well as SNOW_CODE from the data provided by the unique ecosystem data table
}
