% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vri_bem.R
\name{merge_vri_bem}
\alias{merge_vri_bem}
\title{Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory)}
\usage{
merge_vri_bem(vri, bem)
}
\arguments{
\item{vri}{sf object that represent BEM (broad ecosystem mapping) features}

\item{bem}{sf object that represent VRI (vegetation ressource inventory) features}
}
\value{
sf object that represent the original VRI with merged BEM attributes based on largest overlay
}
\description{
Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory)
}
