% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vri_on_bem.R
\name{merge_vri_on_bem}
\alias{merge_vri_on_bem}
\title{Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory) features}
\usage{
merge_vri_on_bem(vri, bem, return_intersection_dt = FALSE)
}
\arguments{
\item{vri}{sf object that represent VRI (vegetation ressource inventory) features}

\item{bem}{sf object that represent BEM (broad ecosystem mapping) features}

\item{return_intersection_dt}{boolean, if TRUE will return a list that contains the sf object and the intesection data.table of VRI and BEM}
}
\value{
sf object that represent the original VRI with merged BEM attributes based on largest overlay
}
\description{
Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory) features
}
