% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_wetland.R
\name{merge_wetlands}
\alias{merge_wetlands}
\title{Merge wetlands into VRI-BEM}
\usage{
merge_wetlands(vri_bem, wetlands)
}
\arguments{
\item{vri_bem}{sf object that represents the VRI-BEM feature class}

\item{wetlands}{sf object that represents Wetlands polygon feature class}
}
\value{
data.table object
}
\description{
Merge wetlands into VRI-BEM
}
