% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_raster_shelter_disturbance.R
\name{pa_Shelter_EWH}
\alias{pa_Shelter_EWH}
\alias{pa_Shelter_WFD}
\alias{pa_dash_security}
\title{Effective Winter Habitat with or without disturbance}
\usage{
pa_Shelter_EWH(shelter, disturbance, dist = units::as_units("400 m"))

pa_Shelter_WFD(
  shelter,
  static_forage = shelter,
  AOI_WMUs_6,
  dist = units::as_units("500 m")
)

pa_dash_security(shelter, AOI_WMUs_6, dist = units::as_units("100 m"))
}
\arguments{
\item{shelter}{vri-bem sf object.}

\item{disturbance}{sf object}

\item{dist}{unit}

\item{static_forage}{sf object}

\item{AOI_WMUs_6}{sf object}

\item{rasterize}{A logical. Should the results be rasterized.}
}
\description{
Effective Winter Habitat with or without disturbance

Static forage within distance from shelters

Dash security distance
}
