% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_rivers}
\alias{rasterize_rivers}
\title{Convert rivers to raster}
\usage{
rasterize_rivers(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  numeric_attributes = NULL,
  character_attributes = NULL,
  date_attributes = NULL,
  burn = "rivers",
  output_raster = FALSE,
  verbose = TRUE
)
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Create one raster layers for rivers with value 1 when there is rivers and 0 otherwise
}
