% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_rivers_materialized}
\alias{rasterize_rivers_materialized}
\title{Convert rivers to raster using materialized on-disk coded fields}
\usage{
rasterize_rivers_materialized(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  aoi = NULL,
  numeric_attributes = NULL,
  character_attributes = NULL,
  date_attributes = NULL,
  burn = "rivers",
  output_raster = FALSE,
  verbose = TRUE,
  dynamic_factor_conv = FALSE,
  dynamic_factor_strategy = c("append", "replace")
)
}
\arguments{
\item{src_datasource}{character, path of source file}

\item{dst_filename}{character, path of destination file}

\item{layer}{character, name of the layer of interest}

\item{reference}{character of SpatRaster, will be used as reference raster for extent, resolution and crs}

\item{numeric_attributes}{character vector of names of attributes of type numeric}

\item{character_attributes}{character vector of names of attributes of type character}

\item{date_attributes}{character vector of names of attributes of type date}

\item{burn}{character, create a layer with the name of the "burn" argument which burn the value 1 where polygons intersect with the extent of the raster}

\item{output_raster}{boolean, if TRUE, the resulting raster will be returned}

\item{verbose}{boolean, if TRUE progression message will be printed}

\item{dynamic_factor_conv}{Logical. If TRUE, build character lookup tables
from distinct values present in the source data instead of relying only on
package lookup data.}

\item{dynamic_factor_strategy}{Character. Either "append" (preserve existing
codes and append new values) or "replace" (rebuild codes from observed
values only).}
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert rivers to raster using materialized on-disk coded fields
}
