% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_sf.R, R/utils.R
\name{rasterize_sf}
\alias{rasterize_sf}
\title{Convert sf to raster}
\usage{
rasterize_sf(
  x,
  crs = albers,
  resolution = units::as_units("100 m"),
  extent = terra::ext(159587.5, 1881187.5, 173787.5, 1748187.5),
  field = 1,
  background = 0,
  ...
)

rasterize_sf(
  x,
  crs = albers,
  resolution = units::as_units("100 m"),
  extent = terra::ext(159587.5, 1881187.5, 173787.5, 1748187.5),
  field = 1,
  background = 0,
  ...
)
}
\arguments{
\item{x}{sf object}

\item{crs}{Target crs}

\item{resolution}{Target resolution as unit.}

\item{extent}{Target raster extent.}

\item{field}{character or numeric. If field is a character, it should a variable name in \code{x}.
If field is numeric it typically is a single number or a vector of length \code{nrow(x)}.
The values are recycled to \code{nrow(x)}.}

\item{background}{numeric. Value to put in the cells that are not covered by any
of the features of \code{x}. Default is \code{NA}.}

\item{...}{Not used.}

\item{src_datasource}{character, path of source file}

\item{dst_filename}{character, path of destination file}

\item{layer}{character, name of the layer of interest}

\item{reference}{character of SpatRaster, will be used as reference raster for extent, resolution and crs}

\item{numeric_attributes}{character vector of names of attributes of type numeric}

\item{character_attributes}{character vector of names of attributes of type character}

\item{date_attributes}{character vector of names of attributes of type date}

\item{factor_conv_list}{list of data.table that contains corresponding factor values for each variable that is not numeric}

\item{burn}{character, create a layer with the name of the "burn" argument which burn the value 1 where polygons intersect with the extent of the raster}

\item{output_raster}{boolean, if TRUE, the resulting raster will be returned}

\item{verbose}{boolean, if TRUE progression message will be printed}

\item{a_srs}{character, assign output SRS (WKT definition)}

\item{te}{numeric vector, target extent: (xmin ymin xmax ymax)}

\item{tr}{numeric vector, target resolution: (xres yres)}
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert all attributes of interest of sf into raster layers
}
