% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_sf_gdal}
\alias{rasterize_sf_gdal}
\title{Convert sf to raster}
\usage{
rasterize_sf_gdal(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  numeric_attributes = NULL,
  character_attributes = NULL,
  date_attributes = NULL,
  factor_conv_list = NULL,
  burn = NULL,
  output_raster = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{src_datasource}{character, path of source file}

\item{dst_filename}{character, path of destination file}

\item{layer}{character, name of the layer of interest}

\item{reference}{character of SpatRaster, will be used as reference raster for extent, resolution and crs}

\item{numeric_attributes}{character vector of names of attributes of type numeric}

\item{character_attributes}{character vector of names of attributes of type character}

\item{date_attributes}{character vector of names of attributes of type date}

\item{factor_conv_list}{list of data.table that contains corresponding factor values for each variable that is not numeric}

\item{burn}{character, create a layer with the name of the "burn" argument which burn the value 1 where polygons intersect with the extent of the raster}

\item{output_raster}{boolean, if TRUE, the resulting raster will be returned}

\item{verbose}{boolean, if TRUE progression message will be printed}
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert all attributes of interest of sf into raster layers
}
