% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_vri}
\alias{rasterize_vri}
\title{Convert vri to raster}
\usage{
rasterize_vri(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  numeric_attributes = c(paste0("SPEC_PCT_", 1:6), "CR_CLOSURE", "COV_PCT_1",
    "PROJ_AGE_1"),
  character_attributes = c(paste0("BCLCS_LV_", 1:5), paste0("SPEC_CD_", 1:6),
    "LAND_CD_1", "LBL_VEGCOV"),
  date_attributes = "HRVSTDT",
  burn = NULL,
  output_raster = FALSE,
  verbose = TRUE
)
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert all attributes of interest of vri into raster layers
}
