% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_vri_materialized}
\alias{rasterize_vri_materialized}
\title{Convert vri to raster using materialized on-disk coded fields}
\usage{
rasterize_vri_materialized(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  aoi = NULL,
  numeric_attributes = c(paste0("SPEC_PCT_", 1:6), "CR_CLOSURE", "COV_PCT_1",
    "PROJ_AGE_1"),
  character_attributes = c(paste0("BCLCS_LV_", 1:5), paste0("SPEC_CD_", 1:6),
    "LAND_CD_1", "LBL_VEGCOV"),
  date_attributes = "HRVSTDT",
  burn = NULL,
  output_raster = FALSE,
  verbose = TRUE,
  dynamic_factor_conv = FALSE,
  dynamic_factor_strategy = c("append", "replace")
)
}
\arguments{
\item{src_datasource}{character, path of source file}

\item{dst_filename}{character, path of destination file}

\item{layer}{character, name of the layer of interest}

\item{reference}{character of SpatRaster, will be used as reference raster for extent, resolution and crs}

\item{numeric_attributes}{character vector of names of attributes of type numeric}

\item{character_attributes}{character vector of names of attributes of type character}

\item{date_attributes}{character vector of names of attributes of type date}

\item{burn}{character, create a layer with the name of the "burn" argument which burn the value 1 where polygons intersect with the extent of the raster}

\item{output_raster}{boolean, if TRUE, the resulting raster will be returned}

\item{verbose}{boolean, if TRUE progression message will be printed}

\item{dynamic_factor_conv}{Logical. If TRUE, build character lookup tables
from distinct values present in the source data instead of relying only on
package lookup data.}

\item{dynamic_factor_strategy}{Character. Either "append" (preserve existing
codes and append new values) or "replace" (rebuild codes from observed
values only).}
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert vri to raster using materialized on-disk coded fields
}
