% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_wetlands}
\alias{rasterize_wetlands}
\title{Convert wetlands to raster}
\usage{
rasterize_wetlands(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  numeric_attributes = NULL,
  character_attributes = NULL,
  date_attributes = NULL,
  burn = "wl_pct",
  output_raster = FALSE,
  verbose = TRUE
)
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Create one raster layers for wetlands with value 1 when there is wetlands and 0 otherwise
}
