% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_layers.R
\name{read_bem}
\alias{read_bem}
\title{Read BEM spatial dataset}
\usage{
read_bem(dsn, layer = "BEM")
}
\arguments{
\item{dsn}{data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder, or contain the name and access credentials of a database); in case of GeoJSON, dsn may be the character string holding the geojson data. It can also be an open database connection.}

\item{layer}{layer name (varies by driver, may be a file name without extension); in case layer is missing, st_read will read the first layer of dsn, give a warning and (unless quiet = TRUE) print a message when there are multiple layers, or give an error if there are no layers in dsn. If dsn is a database connection, then layer can be a table name or a database identifier (see Id). It is also possible to omit layer and rather use the query argument.}
}
\value{
sf object
}
\description{
Read the broad ecosystem mapping (BEM) layer
}
