% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_layers.R
\name{read_tsa}
\alias{read_tsa}
\title{Read aoi (TSAs)}
\usage{
read_tsa(tsa_name, Skeena_boundary = TRUE)
}
\arguments{
\item{tsa_name}{Character vector. A vector of TSA_NUMBER_DESCRIPTION values to filter.}

\item{Skeena_boundary}{Boolean. Restrict tsa to Skeena region.}
}
\value{
sf object
}
\description{
Read in the TSA of interest as an aoi
Filter to Skeena region (for TSAs that extend past Skeena region boundary). If not required, Skeena_boundary=FALSE
}
