% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unique_ecosystem_dt.R
\name{read_unique_ecosystem_dt}
\alias{read_unique_ecosystem_dt}
\title{Read and format the unique ecosystem dt}
\usage{
read_unique_ecosystem_dt(file)
}
\arguments{
\item{file}{File name in working directory, path to file (passed through path.expand for convenience), or a URL starting http://, file://, etc. Compressed files with extension ‘.gz’ and ‘.bz2’ are supported if the R.utils package is installed.}
}
\value{
data.table object
}
\description{
Read and format the unique ecosystem dt
}
