% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geometry.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename geometry attribute for sf object}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{sf object}

\item{name}{name of new sf column}
}
\value{
sf object
}
\description{
Rename geometry attribute for sf object
}
