% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrm_calc_ratings.R
\name{rrm_calc_ratings}
\alias{rrm_calc_ratings}
\alias{rrm_missing_lines}
\title{Calculate WHR ratings based on RRM}
\usage{
rrm_calc_ratings(
  rrm_input,
  template,
  rsi_source,
  output = c("value/rating", "full")
)

rrm_missing_lines(x)
}
\arguments{
\item{rrm_input}{A data.table.}

\item{template}{A character. Path to an RRM Excel file.}

\item{rsi_source}{A data.table.}

\item{output}{A character. Either \code{value/rating} or \code{full} for the full
details with all variables from the formula sheet.}

\item{x}{Result of \code{rrm_calc_ratings}.}
}
\value{
A data.table matching RRT sheet with two additional columns for each
formula. Attribute \code{missing_lines} contains a list of missing lines for each
component sheet.

A list of missing lines by sheet or \code{NULL} if none.
}
\description{
Calculate WHR ratings based on RRM

Extract missing lines from the result of \code{rrm_calc_ratings}.
}
\details{
Calculates WHR ratings based on an RRM Excel file
that contains RSI, AVE and RRT worksheets in their expected
formats. Worksheet names must contain prefix RSI_, AVE_ or RRT_.
The RRT_ worksheet will be modified with added columns containing
the calculated ratings.
}
