% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beu_corrections.R
\name{SDEC_adjustment}
\alias{SDEC_adjustment}
\title{SDEC fix}
\usage{
SDEC_adjustment(vri_bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}
}
\description{
Adjust SDEC so all deciles total to 10.
}
\details{
If SDEC_total > 10, subtract from primary ecosystem (BEUMC_S1) until they equal 10
If SDEC_total < 10, add to primary ecosystem until total is 10
}
