% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_shifted_eco_variables.R
\name{set_shifted_eco_variables}
\alias{set_shifted_eco_variables}
\title{Shift eco variables (data.table in-memory version)}
\usage{
set_shifted_eco_variables(
  input_dt,
  i,
  shift_pattern,
  integer_variables_1 = c("TREE_C1", "SHRUB_C1"),
  character_variables_1 = c("BEUMC_S1", "REALM_1", "GROUP_1", "CLASS_1", "KIND_1",
    "SITE_S1", "SITEAM_S1A", "SITEAM_S1B", "SITEAM_S1C", "SITEAM_S1D", "SITEMC_S1",
    "SITE_M1A", "SITE_M1B", "STRCT_S1", "STRCT_M1", "STAND_A1", "SERAL_1", "DISTCLS_1",
    "DISTSCLS_1", "DISSSCLS_1", "SECL_1", "SESUBCL_1", "COND_1", "VIAB_1", "FORESTED_1")
)
}
\arguments{
\item{input_dt}{data.table input data to be modified by reference}

\item{i}{numeric vector that identify which lines of the data we want to apply the shifting}

\item{shift_pattern}{list of vectors , each vector represent a shift, the first element of the vector is which variable we want to change and the second is what to use to feed the first variable}

\item{integer_variables_1}{character vector that represent the name of the variables of type integer with index number 1 we want to shift}

\item{character_variables_1}{character vector that represent the name of the variables of type character with index number 1 we want to shift}
}
\description{
Shift eco variables (data.table in-memory version)
}
