% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vri_bem_output.R
\name{simplify_vri_bem_output}
\alias{simplify_vri_bem_output}
\title{Simplify the VRI_BEM_WHR output
Further reduce layers present in VRI-BEM for sharing and distributing, based on similar reductions for site selection}
\usage{
simplify_vri_bem_output(vri_bem)
}
\arguments{
\item{vri_bem}{VRI-BEM feature class}
}
\value{
simplified vri_bem
}
\description{
Consider retaining BEUMC_S2, STRCT_S2, BEUMC_S3, etc?
}
