% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beu_from_rules_dt.R
\name{sum_pct_for_species_in_list}
\alias{sum_pct_for_species_in_list}
\title{sum_pct_for_species_in_list}
\usage{
sum_pct_for_species_in_list(
  tree_list,
  vri_bem = NULL,
  species_var = NULL,
  pct_var = NULL
)
}
\arguments{
\item{tree_list}{string list of trees}

\item{vri_bem}{sf object of vri-bem, optional}

\item{species_var}{character vector of names of species variables to use in vri bem}

\item{pct_var}{character vector of names of percentage variables to use in vri bem}
}
\value{
string
}
\description{
create string expression that sum the percentage of the species in the list across all pairs of species/pct variables
}
\keyword{internal}
