% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_unique_ecosystem.R
\name{summarize_unique_ecosystem}
\alias{summarize_unique_ecosystem}
\title{Create summary data of unique ecosystem}
\usage{
summarize_unique_ecosystem(vri_bem_dt)
}
\arguments{
\item{vri_bem_dt}{data.table object that represent the vri bem}
}
\value{
data.table that contains the frequency of each unique ecosystem
}
\description{
Generates a data of unique BGC label and habitat combinations for the Look up Table starting point
}
