% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_vri.R,
%   R/update_bem_from_vri_raster.R
\name{update_bem_from_vri}
\alias{update_bem_from_vri}
\title{Update BEM attributes based on VRI attributes}
\usage{
update_bem_from_vri(
  vri_bem,
  rivers,
  beu_bec,
  clear_site_ma = TRUE,
  use_ifelse = TRUE,
  raster_res = NULL
)

update_bem_from_vri(
  vri_bem,
  rivers,
  beu_bec,
  clear_site_ma = TRUE,
  use_ifelse = TRUE,
  raster_res = NULL
)
}
\arguments{
\item{vri_bem}{sf object that represent the combined vri & bem polygon feature class}

\item{rivers}{sf object that represent Rivers polygon feature class (FWA_Rivers)}

\item{beu_bec}{data.table object of allowed BEC and BEM Code Combos}

\item{clear_site_ma}{boolean, if TRUE variable SITE_M1A, SITE_M2A will be cleared}

\item{use_ifelse}{boolean, if TRUE correction done after the combine_duplicated_BEUMC will only be applied on rows that were not affected by the correction of duplicated BEUMC}

\item{raster_res}{numeric vector of resolution example c(25,25)}
}
\value{
sf object which contains adjusted map codes

sf object which contains adjusted map codes
}
\description{
This function updates BEM (broad ecosystem mapping) attributes based on VRI (vegetation resource inventory) attributes, according to the document "Corrections to the BEM map codes".

This function updates BEM (broad ecosystem mapping) attributes based on VRI (vegetation resource inventory) attributes, according to the document "Corrections to the BEM map codes".
}
\details{
This function performs some adjustments to BEM attributes based on VRI attributes :
\itemize{
\item Combine together components with duplicated BEUMC
\item Perform corrections based on BC Land Cover Classification Scheme (BCLCS) and land : assign related SDEC_1 to 10
\item Remove inadequate wetlands
\item Update STAND_... based on species
}

This function performs some adjustments to BEM attributes based on VRI attributes :
\itemize{
\item Combine together components with duplicated BEUMC
\item Perform corrections based on BC Land Cover Classification Scheme (BCLCS) and land : assign related SDEC_1 to 10
\item Remove inadequate wetlands
\item Update STAND_... based on species
}
}
