% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_required_attributes.R
\name{validate_required_attributes}
\alias{validate_required_attributes}
\title{Validate that required attributes are in the object}
\usage{
validate_required_attributes(ifc, required_attributes)
}
\arguments{
\item{ifc}{input feature class}

\item{required_attributes}{names of required attributes}
}
\value{
error if attribute is missing
}
\description{
Validate that required attributes are in the object
}
