% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beu_from_rules_dt.R
\name{var_contains}
\alias{var_contains}
\title{var_contains}
\usage{
var_contains(var_name, rule)
}
\arguments{
\item{var_name}{string name of variable}

\item{rule}{string of rule}
}
\value{
string
}
\description{
create string expression to check if variable contains the rule
}
\keyword{internal}
