% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beu_from_rules_dt.R
\name{var_equal_value}
\alias{var_equal_value}
\title{var_equal_value}
\usage{
var_equal_value(var_name, rule_value)
}
\arguments{
\item{var_name}{string name of variable}

\item{rule_value}{string value to match}
}
\value{
string
}
\description{
create string expression that check if the variable equal a value
}
\keyword{internal}
