% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beu_from_rule_dt.R
\name{var_in_list}
\alias{var_in_list}
\title{var_in_list}
\usage{
var_in_list(var_name, rule_list)
}
\arguments{
\item{var_name}{string name of variable}

\item{rule_list}{string list of values}
}
\value{
string
}
\description{
create string expression that check if the variable is in the list
}
\keyword{internal}
