% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vribem_view.R
\name{vribem_view}
\alias{vribem_view}
\title{Create VRIBEM view}
\usage{
vribem_view(conn, validate_intersect = FALSE)
}
\arguments{
\item{conn}{A duckdb connection.}

\item{validate_intersect}{Boolean, display a message for non intersecting VRI and BEM shape.}
}
\description{
Create VRIBEM view
}
\details{
This function will create view \code{V_VRIBEM} and \code{V_VRIBEM_INTERSECTION} on \code{conn}.
It will contain all VRI and BEM fields.
BGC fields from VRI will have priority over matching BEM fields.
}
