% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_forest_age_class.R
\name{calc_forest_age_class}
\alias{calc_forest_age_class}
\title{Compute variables for structural stage look up and stand composition look up}
\usage{
calc_forest_age_class(vri_bem, most_recent_harvest_year)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation resource inventory) features}

\item{most_recent_harvest_year}{integer that represent the most recent harvest year}
}
\value{
sf object vri
}
\description{
Compute variables for structural stage look up and stand composition look up
}
