% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_vri.R
\name{correct_bem_from_vri}
\alias{correct_bem_from_vri}
\title{Correct BEM attributes based on VRI attributes}
\usage{
correct_bem_from_vri(
  vri_bem,
  beu_bec,
  clear_site_ma = TRUE,
  use_ifelse = TRUE,
  raster_res = NULL
)
}
\arguments{
\item{vri_bem}{sf object that represent the combined vri & bem polygon feature class}

\item{beu_bec}{data.table object of allowed BEC and BEM Code Combos}

\item{clear_site_ma}{boolean, if TRUE variable SITE_M1A, SITE_M2A will be cleared}

\item{use_ifelse}{boolean, if TRUE correction done after the combine_duplicated_BEUMC will only be applied on rows that were not affected by the correction of duplicated BEUMC}

\item{raster_res}{numeric vector of resolution example c(25,25)}
}
\value{
sf object which contains adjusted map codes
}
\description{
Correct BEM attributes based on VRI attributes
}
\details{
This function performs some adjustments to BEM attributes based on VRI attributes :
\itemize{
\item Combine together components with duplicated BEUMC
\item Perform corrections based on BC Land Cover Classification Scheme (BCLCS) and land : assign related SDEC_1 to 10
\item Remove inadequate wetlands
\item Update STAND_... based on species
}
}
