% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hem_rasters.R
\name{create_hem_rasters}
\alias{create_hem_rasters}
\title{Create raster with layers for each hem fields}
\usage{
create_hem_rasters(
  vri_bem,
  path,
  xmin = 1020387.5,
  xmax = 1030387.5,
  ymin = 960987.5,
  ymax = 970987.5,
  resolution = 100,
  crs = albers
)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{path}{character string that represent the path the output raster will be save to}

\item{xmin}{numeric that represent the xmin of the extent of the raster}

\item{xmax}{numeric that represent the xmax of the extent of the raster}

\item{ymin}{numeric that represent the ymin of the extent of the raster}

\item{ymax}{numeric that represent the ymax of the extent of the raster}

\item{resolution}{numeric that represent the resolution of the raster}

\item{crs}{character string of the WKT to be used for the projection of the raster}
}
\value{
spatRaster object
}
\description{
Create raster with layers for each hem fields
}
