% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RRM_ecosystem.R
\name{create_RRM_ecosystem}
\alias{create_RRM_ecosystem}
\title{Create RRM Ecosystem information}
\usage{
create_RRM_ecosystem(bfc)
}
\arguments{
\item{bfc}{bem feature class}
}
\value{
Summary of Area by unique ecosystem
}
\description{
This function generates the unique list of ecosystems and determines all possible combinations of STS and STAND codes
that could apply to the ecosystem at any age value, and expand to include all existing and "potential" unique ecosystems
}
