% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_crown_area_dominant_values.R
\name{find_crown_area_dominant_values}
\alias{find_crown_area_dominant_values}
\title{Find Area-Dominant CROWN_BEAR and CROWN_MOOSE}
\usage{
find_crown_area_dominant_values(vri, bem, intersection_dt = NULL)
}
\arguments{
\item{vri}{sf object that represent VRI (vegetation ressource inventory) features}

\item{bem}{sf object that represent BEM (broad ecosystem mapping) features}

\item{intersection_dt}{data.table  area and index of vri and bem intersections}
}
\value{
sf object
}
\description{
This function overlays a BEM and VRI feature classes and determines the area-dominant CROWN_BEAR and CROWN_MOOSE values for each BEM polygon.
}
