% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_db.R
\name{init_db}
\alias{init_db}
\alias{init_vri}
\alias{init_bem}
\alias{init_wetlands}
\alias{init_rivers}
\alias{init_lakes}
\alias{init_glaciers}
\alias{init_ccb}
\alias{init_burn}
\alias{init_fire}
\alias{init_tsa}
\alias{init_conn}
\title{Initialize a database}
\usage{
init_db(dbdir = defdb(), ask = interactive(), bem_dsn = NULL, prem_dsn = NULL)

init_vri(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "VEG_COMP_LYR_R1_POLY",
  geom = "Shape"
)

init_bem(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "BEM",
  geom = "geom"
)

init_wetlands(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "FWA_WETLANDS_POLY",
  geom = "geom"
)

init_rivers(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "FWA_RIVERS_POLY",
  geom = "geom"
)

init_lakes(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "FWA_LAKES_POLY",
  geom = "geom"
)

init_glaciers(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "BTM_PLU_V1",
  geom = "geom"
)

init_ccb(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "Cut_Block_all_BC",
  geom = "geom"
)

init_burn(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "WHSE_FOREST_VEGETATION_VEG_BURN_SEVERITY_SP",
  geom = "geom"
)

init_fire(
  conn = init_conn(),
  ask = interactive(),
  dsn = NULL,
  layer = "WHSE_LAND_AND_NATURAL_RESOURCE.PROT_HISTORICAL_FIRE_POLYS_SP",
  geom = "geom"
)

init_tsa(
  conn = init_conn(),
  ask = interactive(),
  tsa_name = c(),
  Skeena_boundary = TRUE,
  tsa_dsn = NULL
)

init_conn(dbdir = defdb())
}
\arguments{
\item{dbdir}{Location for database files. Should be a path to an existing
directory in the file system. With the default (or ""), all data is kept
in RAM.}

\item{ask}{Boolean, whether to ask before re-initializing existing tables.}

\item{bem_dsn}{data source name for BEM data.}

\item{conn}{Connection to database.}

\item{dsn}{data source name (interpretation varies by driver - for some
drivers, dsn is a file name, but may also be a folder, or contain the name
and access credentials of a database); in case of GeoJSON, dsn may be the
character string holding the geojson data. It can also be an open database
connection. When dsn is left empty/NULL the polygons will be fetched from
the BC Data Catalog.}

\item{layer}{layer name (varies by driver, may be a file name without
extension); in case layer is missing, st_read will read the first layer of
dsn, give a warning and (unless quiet = TRUE) print a message when there
are multiple layers, or give an error if there are no layers in dsn. If dsn
is a database connection, then layer can be a table name or a database
identifier (see Id). It is also possible to omit layer and rather use the
query argument.}

\item{geom}{geometry field name to use. Default to \code{Shape}.}

\item{tsa_name}{Character vector.
A vector of TSA_NUMBER_DESCRIPTION values to filter.}

\item{Skeena_boundary}{Boolean. Restrict tsa to Skeena region.}
}
\description{
Initialize a database

Initialize and load spatial dataset to database
}
\details{
This needs to be run once to create the database and load the
required spatial dataset into it. The database can be stored locally
for faster processing.
}
