% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vri_on_bem.R
\name{merge_bem_on_vri}
\alias{merge_bem_on_vri}
\title{Merge BEM (broad ecosystem mapping) attributes on VRI (vegetation ressource inventory) features}
\usage{
merge_bem_on_vri(vri, bem, return_intersection_dt = FALSE)
}
\arguments{
\item{vri}{sf object that represent VRI (vegetation ressource inventory) features}

\item{bem}{sf object that represent BEM (broad ecosystem mapping) features}

\item{return_intersection_dt}{boolean, if TRUE will return a list that contains the sf object and the intesection data.table of VRI and BEM}
}
\value{
sf object that represent the original VRI with merged BEM attributes based on largest overlay.
The output will be a copy of the VRI polygons clipped to the BEM, then exploded to singlepart, and its slivers will be eliminated.
Then its attribute table will have all of the BEM attribute fields added to the existing VRI attributes.
The BEI attribute fields will be populated by copying the BEI values of the majority area BEI polygon within each VRI polygon.
}
\description{
This function copies all of the attributes in the specified BEM to each polygon in the specified VRI.
}
