% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rrm_on_vri.R
\name{merge_rrm_on_vri}
\alias{merge_rrm_on_vri}
\title{merge_rrm_on_vri}
\usage{
merge_rrm_on_vri(vri_bem, rrm_dt)
}
\arguments{
\item{vri_bem}{sf object that represent VRI (vegetation ressource inventory) features}

\item{rrm_dt}{data.table object that contains the rrm}
}
\value{
sf  vri-bem object with new columns for rating
}
\description{
merge suitability and capability rating from rrm onto vri-bem
and calculate highest value and weighted average for each scores
}
