% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_bem_materialized}
\alias{rasterize_bem_materialized}
\title{Convert Broad ecosystem mapping to raster using materialized on-disk coded fields}
\usage{
rasterize_bem_materialized(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  aoi = NULL,
  numeric_attributes = c("TEIS_ID", "Area_Ha", "AGE_CL_STS", "SDEC_1", "SDEC_2",
    "SDEC_3", "BGC_VRT"),
  character_attributes = c("BGC_ZONE", "BGC_SUBZON", "BGC_PHASE", "SLOPE_MOD", "ECO_SEC",
    "BEUMC_S1", "REALM_1", "GROUP_1", "CLASS_1", "KIND_1", "SITE_S1", "SITEAM_S1A",
    "SITEAM_S1B", "SITEAM_S1C", "SITEAM_S1D", "SITEMC_S1", "SITE_M1A", "SITE_M1B",
    "STRCT_S1", "STRCT_M1", "STAND_A1", "SERAL_1", "DISTCLS_1", "DISTSCLS_1",
    "DISSSCLS_1", "SECL_1", "SESUBCL_1", "COND_1", "VIAB_1", "FORESTED_1", "TREE_C1",
    "SHRUB_C1", "BEUMC_S2", "REALM_2", "GROUP_2", "CLASS_2", "KIND_2", "SITE_S2",
    "SITEAM_S2A", "SITEAM_S2B", "SITEAM_S2C", "SITEAM_S2D", 
     "SITEMC_S2",
    "SITE_M2A", "SITE_M2B", "STRCT_S2", "STRCT_M2", "STAND_A2", "SERAL_2", "DISTCLS_2",
    "DISTSCLS_2", "DISSSCLS_2", "SECL_2", "SESUBCL_2", "COND_2", "VIAB_2", "FORESTED_2",
    "TREE_C2", "SHRUB_C2", "BEUMC_S3", "REALM_3", "GROUP_3", "CLASS_3", "KIND_3",
    "SITE_S3", "SITEAM_S3A", "SITEAM_S3B", "SITEAM_S3C", "SITEAM_S3D", "SITEMC_S3",
    "SITE_M3A", "SITE_M3B", "STRCT_S3", "STRCT_M3", "STAND_A3", "SERAL_3", "DISTCLS_3",
    "DISTSCLS_3", "DISSSCLS_3", "SECL_3", "SESUBCL_3", "COND_3", "VIAB_3", "FORESTED_3",
    
     "TREE_C3", "SHRUB_C3"),
  date_attributes = NULL,
  burn = NULL,
  output_raster = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{src_datasource}{character, path of source file}

\item{dst_filename}{character, path of destination file}

\item{layer}{character, name of the layer of interest}

\item{reference}{character of SpatRaster, will be used as reference raster for extent, resolution and crs}

\item{numeric_attributes}{character vector of names of attributes of type numeric}

\item{character_attributes}{character vector of names of attributes of type character}

\item{date_attributes}{character vector of names of attributes of type date}

\item{burn}{character, create a layer with the name of the "burn" argument which burn the value 1 where polygons intersect with the extent of the raster}

\item{output_raster}{boolean, if TRUE, the resulting raster will be returned}

\item{verbose}{boolean, if TRUE progression message will be printed}
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Convert Broad ecosystem mapping to raster using materialized on-disk coded fields
}
