% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raterize_sf_gdal.R
\name{rasterize_ccb}
\alias{rasterize_ccb}
\title{Convert consolidated cutblocks to raster}
\usage{
rasterize_ccb(
  src_datasource,
  dst_filename,
  layer = NULL,
  a_srs = NULL,
  te = NULL,
  tr = NULL,
  reference = NULL,
  numeric_attributes = "HARVESTYR",
  character_attributes = NULL,
  date_attributes = NULL,
  burn = NULL,
  output_raster = FALSE,
  verbose = TRUE
)
}
\value{
SpatRaster if output_raster is TRUE, NULL otherwise
}
\description{
Create one raster layers for consolidated cutblocks with value from HARVEST_YEAR
}
