% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rasterize_sf}
\alias{rasterize_sf}
\title{Rasterize an sf object}
\usage{
rasterize_sf(
  x,
  crs = albers,
  resolution = units::as_units("100 m"),
  extent = terra::ext(159587.5, 1881187.5, 173787.5, 1748187.5),
  field = 1,
  background = 0,
  ...
)
}
\arguments{
\item{x}{sf object}

\item{crs}{Target crs}

\item{resolution}{Target resolution as unit.}

\item{extent}{Target raster extent.}

\item{field}{character or numeric. If field is a character, it should a variable name in \code{x}.
If field is numeric it typically is a single number or a vector of length \code{nrow(x)}.
The values are recycled to \code{nrow(x)}.}

\item{background}{numeric. Value to put in the cells that are not covered by any
of the features of \code{x}. Default is \code{NA}.}

\item{...}{Not used.}
}
\description{
Rasterize an sf object
}
