% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_vri.R
\name{update_bem_from_vri}
\alias{update_bem_from_vri}
\title{Upadate BEM (broad ecosystem mapping) attributes based on VRI (vegetation resource inventory) attributes}
\usage{
update_bem_from_vri(ifc, rfc, clear_site_ma = TRUE)
}
\arguments{
\item{ifc}{sf object that represent the input polygon feature class}

\item{rfc}{sf object that represent Rivers polygon feature class (FWA_Rivers)}

\item{clear_site_ma}{boolean, if TRUE variable SITE_M1A, SITE_M2A will be cleared}

\item{bec_bea}{data.table object of allowed BEC and BEM Code Combos}
}
\value{
sf object
}
\description{
Upadate BEM (broad ecosystem mapping) attributes based on VRI (vegetation resource inventory) attributes
}
