% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bem_from_wet.R
\name{update_bem_from_wet}
\alias{update_bem_from_wet}
\title{Upadate BEM (broad ecosystem mapping) attributes based on Wetland polygons attributes}
\usage{
update_bem_from_wet(bfc, wfc, buc)
}
\arguments{
\item{bfc}{sf object that represents the BEM feature class}

\item{wfc}{sf object that represents Wetlands polygon feature class}

\item{buc}{data.table object that represents BEU wetland update table}
}
\value{
sf object
}
\description{
Upadate BEM (broad ecosystem mapping) attributes based on Wetland polygons attributes
}
